Seem to have fixed the height issue?
Plz no judgement for how late this submission is :P

I have fixed the height issue, not the most elegant solution but it seems to work until we get down to the height of the sidebar which then become the limiting factor.

We also need to do something about the width, when the page isn't as wide as the contents of the navbar then the page introduces a scroll bar.