Finished About, Contact Form, started featured page with social media link
I couldn't sleep so I coded.

Fixed the alignment issue with image link on about page.

Made a contact form for businesses and embedded it in the business contact page.

Created our very own personal instagram hashtag and linked to a feed of this on the featured page (using my own images). May be worth adding a carousel of recent images to the page to bring it to life? Will look at when it isn't 3am ;)

NEED TO DO:
Add the following locations to the map (need to do in order for my instagram images to be of any use):
- Byron Burger, Gloucester Road
- Pain Quotidien, South Ken
- Muriel's kitchen, South Ken
- Leon, High Street Ken
- Hummingbird Bakery, South Ken